refactor: rename wemixgov to systemcontracts and clean up code
