+++ Makefile
@@ -8,7 +8,7 @@ GOBIN = ./build/bin
 GO ?= latest
 GORUN = go run
 
-SUCCESS_TESTS = github.com/ethereum/go-ethereum/accounts github.com/ethereum/go-ethereum/accounts/abi github.com/ethereum/go-ethereum/cmd/abigen github.com/ethereum/go-ethereum/cmd/clef github.com/ethereum/go-ethereum/cmd/evm/... github.com/ethereum/go-ethereum/cmd/devp2p/...  github.com/ethereum/go-ethereum/cmd/ethkey github.com/ethereum/go-ethereum/cmd/rlpdump github.com/ethereum/go-ethereum/common/... github.com/ethereum/go-ethereum/core/... github.com/ethereum/go-ethereum/ethclient/... github.com/ethereum/go-ethereum/consensus/misc/...  github.com/ethereum/go-ethereum/consensus/wbft/...  github.com/ethereum/go-ethereum/crypto/...  github.com/ethereum/go-ethereum/ethdb/...  github.com/ethereum/go-ethereum/ethstats github.com/ethereum/go-ethereum/event github.com/ethereum/go-ethereum/log github.com/ethereum/go-ethereum/metrics/...  github.com/ethereum/go-ethereum/miner github.com/ethereum/go-ethereum/node github.com/ethereum/go-ethereum/p2p/...  github.com/ethereum/go-ethereum/params github.com/ethereum/go-ethereum/rlp/...  github.com/ethereum/go-ethereum/signer/...  github.com/ethereum/go-ethereum/tests/fuzzers/...  github.com/ethereum/go-ethereum/trie/...  github.com/ethereum/go-ethereum/triedb/...  github.com/ethereum/go-ethereum/graphql/...  github.com/ethereum/go-ethereum/internal/...  github.com/ethereum/go-ethereum/rpc github.com/ethereum/go-ethereum/cmd/geth github.com/ethereum/go-ethereum/cmd/utils github.com/ethereum/go-ethereum/console github.com/ethereum/go-ethereum/eth/... github.com/ethereum/go-ethereum/wemixgov/...
+SUCCESS_TESTS = github.com/ethereum/go-ethereum/accounts github.com/ethereum/go-ethereum/accounts/abi github.com/ethereum/go-ethereum/cmd/abigen github.com/ethereum/go-ethereum/cmd/clef github.com/ethereum/go-ethereum/cmd/evm/... github.com/ethereum/go-ethereum/cmd/devp2p/...  github.com/ethereum/go-ethereum/cmd/ethkey github.com/ethereum/go-ethereum/cmd/rlpdump github.com/ethereum/go-ethereum/common/... github.com/ethereum/go-ethereum/core/... github.com/ethereum/go-ethereum/ethclient/... github.com/ethereum/go-ethereum/consensus/misc/...  github.com/ethereum/go-ethereum/consensus/wbft/...  github.com/ethereum/go-ethereum/crypto/...  github.com/ethereum/go-ethereum/ethdb/...  github.com/ethereum/go-ethereum/ethstats github.com/ethereum/go-ethereum/event github.com/ethereum/go-ethereum/log github.com/ethereum/go-ethereum/metrics/...  github.com/ethereum/go-ethereum/miner github.com/ethereum/go-ethereum/node github.com/ethereum/go-ethereum/p2p/...  github.com/ethereum/go-ethereum/params github.com/ethereum/go-ethereum/rlp/...  github.com/ethereum/go-ethereum/signer/...  github.com/ethereum/go-ethereum/tests/fuzzers/...  github.com/ethereum/go-ethereum/trie/...  github.com/ethereum/go-ethereum/triedb/...  github.com/ethereum/go-ethereum/graphql/...  github.com/ethereum/go-ethereum/internal/...  github.com/ethereum/go-ethereum/rpc github.com/ethereum/go-ethereum/cmd/geth github.com/ethereum/go-ethereum/cmd/utils github.com/ethereum/go-ethereum/console github.com/ethereum/go-ethereum/eth/... github.com/ethereum/go-ethereum/systemcontracts/...
 
 # TODO: move this cases to SUCCESS_TESTS one by one after making it to be success
 FAILURE_TESTS = github.com/ethereum/go-ethereum/tests
