+++ consensus/wbft/engine/engine_test.go
@@ -43,7 +43,7 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/rlp"
-	govwbft "github.com/ethereum/go-ethereum/wemixgov/governance-wbft"
+	"github.com/ethereum/go-ethereum/systemcontracts"
 )
 
 // RLP data can be generated by using TestGeneratingGenesisExtra in testutils package
@@ -422,9 +422,9 @@ func updateChainConfig(chainConfig *params.ChainConfig, newValidators []account)
 		vals += v.addr.Hex()
 		blsKeys += "0x" + hexutil.Encode(v.blsKey.PublicKey().Marshal())[2:]
 	}
-	chainConfig.Anzeon.SystemContracts.GovValidator.Params[govwbft.GOV_BASE_PARAM_MEMBERS] = vals
-	chainConfig.Anzeon.SystemContracts.GovValidator.Params[govwbft.GOV_VALIDATOR_PARAM_VALIDATORS] = vals
-	chainConfig.Anzeon.SystemContracts.GovValidator.Params[govwbft.GOV_VALIDATOR_PARAM_BLS_KEYS] = blsKeys
+	chainConfig.Anzeon.SystemContracts.GovValidator.Params[systemcontracts.GOV_BASE_PARAM_MEMBERS] = vals
+	chainConfig.Anzeon.SystemContracts.GovValidator.Params[systemcontracts.GOV_VALIDATOR_PARAM_VALIDATORS] = vals
+	chainConfig.Anzeon.SystemContracts.GovValidator.Params[systemcontracts.GOV_VALIDATOR_PARAM_BLS_KEYS] = blsKeys
 }
 
 func TestEpochInfo(t *testing.T) {
