+++ systemcontracts/contracts.go
@@ -16,7 +16,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-wemix-wbft library. If not, see <http://www.gnu.org/licenses/>.
 
-package govwbft
+package systemcontracts
 
 import (
 	_ "embed"
@@ -30,7 +30,7 @@ const (
 )
 
 var (
-	//go:embed govcontracts/v1/GovValidator
+	//go:embed artifacts/v1/GovValidator
 	GovValidatorContractV1 string
 
 	SystemContractCodes map[string]map[string]string
