+++ systemcontracts/gov_base_test.go
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the stable-one library. If not, see <http://www.gnu.org/licenses/>.
 
-package govwbft
+package systemcontracts
 
 import (
 	"testing"
