+++ systemcontracts/solidity/abstracts/AbstractFiatToken.sol
