+++ systemcontracts/solidity/interfaces/ICoinManager.sol
