+++ systemcontracts/solidity/interfaces/IMinterManagement.sol
