+++ systemcontracts/solidity/libraries/SignatureChecker.sol
