+++ systemcontracts/solidity/remappings.txt
@@ -1,0 +1,2 @@
+@openzeppelin/contracts/=openzeppelin/contracts/contracts/
+@openzeppelin/contracts-upgradeable/=openzeppelin/contracts-upgradeable/contracts/
