+++ systemcontracts/stateutil.go
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-wemix-wbft library. If not, see <http://www.gnu.org/licenses/>.
 
-package govwbft
+package systemcontracts
 
 import (
 	"math/big"
