+++ systemcontracts/systemcontracts.go
@@ -16,7 +16,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-wemix-wbft library. If not, see <http://www.gnu.org/licenses/>.
 
-package govwbft
+package systemcontracts
 
 import (
 	"github.com/ethereum/go-ethereum/params"
