+++ systemcontracts/test/utils.go
@@ -19,24 +19,19 @@ package test
 
 import (
 	"context"
-	"crypto/ecdsa"
 	"fmt"
 	"math/big"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethclient/simulated"
 	"github.com/ethereum/go-ethereum/params"
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 )
 
@@ -76,16 +71,6 @@ func ExpectedRevert(t *testing.T, err error, args ...interface{}) {
 			require.Equal(t, strings.TrimSpace(message), strings.TrimSpace(errStr))
 		}
 	}
-}
-
-func ToBytes32(str string) [32]byte {
-	bytes := []byte(str)
-	if len(bytes) > 32 {
-		bytes = bytes[:32]
-	}
-	var copied = [32]byte{}
-	copy(copied[:], bytes)
-	return copied
 }
 
 var eoas = make(map[string]*bind.TransactOpts)
@@ -204,23 +189,6 @@ func getNonce(backend interface {
 	}
 }
 
-type EOA struct {
-	PrivateKey *ecdsa.PrivateKey
-	Address    common.Address
-}
-
-type CA struct {
-	Address common.Address
-}
-
-func NewEOA() (eoa *EOA) {
-	pk, _ := crypto.GenerateKey()
-	return &EOA{
-		PrivateKey: pk,
-		Address:    crypto.PubkeyToAddress(pk.PublicKey),
-	}
-}
-
 func NewTxOptsWithValue(t *testing.T, eoa *EOA, value *big.Int) *bind.TransactOpts {
 	opts, err := bind.NewKeyedTransactorWithChainID(eoa.PrivateKey, params.TestWBFTChainConfig.ChainID)
 	require.NoError(t, err)
@@ -230,100 +198,20 @@ func NewTxOptsWithValue(t *testing.T, eoa *EOA, value *big.Int) *bind.TransactOp
 	return opts
 }
 
-// evnet_unpack
-type allEventsType map[string]abi.Event
+func towei(x int64) *big.Int {
+	return new(big.Int).Mul(big.NewInt(x), big.NewInt(params.Ether))
+}
 
-var (
-	allEventsLock = sync.RWMutex{}
-	allEvents     = allEventsType{}
-)
+func toGwei(x int64) *big.Int {
+	return new(big.Int).Mul(big.NewInt(x), big.NewInt(params.GWei))
+}
 
-func collectEvent(abi *abi.ABI) error {
-	for n, e := range abi.Events {
-		func() {
-			allEventsLock.Lock()
-			defer allEventsLock.Unlock()
-			if exist, ok := allEvents[n]; ok {
-				if exist.String() == e.String() {
-					goto skip
-				}
-			}
-			allEvents[n] = e
-		skip:
-		}()
+func ToBytes32(str string) [32]byte {
+	bytes := []byte(str)
+	if len(bytes) > 32 {
+		bytes = bytes[:32]
 	}
-
-	return nil
-}
-
-// error_unpack
-type allErrorsType map[[4]byte]abi.Error
-
-var (
-	allErrorsLock = sync.RWMutex{}
-	allErrors     = allErrorsType{}
-)
-
-func collectErrors(abi *abi.ABI) error {
-	for _, e := range abi.Errors {
-		func() {
-			allErrorsLock.Lock()
-			defer allErrorsLock.Unlock()
-			sig := [4]byte{}
-			copy(sig[:], e.ID[:4])
-			if _, ok := allErrors[sig]; !ok {
-				allErrors[sig] = e
-			}
-		}()
-	}
-
-	return nil
-}
-
-type RevertError struct {
-	ABI    abi.Error
-	Output interface{}
-}
-
-func (r *RevertError) Error() string {
-	return fmt.Sprintf("%s: %s %v", vm.ErrExecutionReverted, r.ABI.Sig, r.Output)
-}
-
-// ErrorCode returns the JSON error code for a revert.
-// See: https://github.com/ethereum/wiki/wiki/JSON-RPC-Error-Codes-Improvement-Proposal
-func NewRevertError(err error) error {
-	if revert, ok := err.(interface {
-		ErrorCode() int
-		ErrorData() interface{}
-	}); !ok || revert.ErrorCode() != 3 {
-		return err
-	} else {
-		if data, ok := revert.ErrorData().(string); !ok {
-			return err
-		} else {
-			datas := hexutil.MustDecode(data)
-			if revertErr, ok := UnpackError(datas); ok {
-				return revertErr
-			} else {
-				reason, errUnpack := abi.UnpackRevert(datas)
-				if errUnpack == nil {
-					return fmt.Errorf("execution reverted: %v", reason)
-				} else {
-					return errors.New("execution reverted")
-				}
-			}
-		}
-	}
-}
-
-func UnpackError(result []byte) (error, bool) {
-	sig := [4]byte{}
-	copy(sig[:], result[:4])
-	if errABI, ok := allErrors[sig]; !ok {
-		return nil, false
-	} else if output, err := errABI.Unpack(result); err != nil {
-		return nil, false
-	} else {
-		return &RevertError{errABI, output}, true
-	}
+	var copied = [32]byte{}
+	copy(copied[:], bytes)
+	return copied
 }
