--- wemixgov/governance-contract/contracts-wbft/.prettierrc
@@ -1,14 +1,0 @@
-{
-  "overrides": [
-    {
-      "files": "*.sol",
-      "options": {
-        "printWidth": 150,
-        "tabWidth": 4,
-        "useTabs": false,
-        "singleQuote": false,
-        "bracketSpacing": true
-      }
-    }
-  ]
-}
